Changed DisplayHost#scale to be int instead of double, since we can get paint artifacts when we use non-integer scale values
git-svn-id: https://svn.apache.org/repos/asf/incubator/pivot/trunk@764608 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/pivot/wtk/Display.java
@@ -81,7 +81,7 @@ public final class Display extends Container {
         if (immediate) {
             Graphics2D graphics = (Graphics2D)displayHost.getGraphics();
 
-            double scale = displayHost.getScale();
+            int scale = displayHost.getScale();
             if (scale != 1) {
                 graphics.scale(scale, scale);
             }
